Some juicy updates

+++ bootstrap.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 CURRENT_DIRECTORY=$(pwd)
 
-ln -s $CURRENT_DIRECTORY/.vimrc ~/.vimrc
-ln -s $CURRENT_DIRECTORY/.tmux.conf ~/.tmux.conf
+if [ ! -f ~/.vimrc ]; then
+    ln -s $CURRENT_DIRECTORY/.vimrc ~/.vimrc
+fi
+
+if [ ! -f ~/.tmux.conf ]; then
+    ln -s $CURRENT_DIRECTORY/.tmux.conf ~/.tmux.conf
+fi
+
+if [ ! -f ~/.gvimrc ]; then
+    ln -s $CURRENT_DIRECTORY/.gvimrc ~/.gvimrc
+fi
